requirements.txt
appdirs==1.4.3
astroid==2.0.4
astropy==3.0.4
atomicwrites==1.2.1
attrs==18.2.0
Automat==0.7.0
Babel==2.6.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
defusedxml==0.5.0
distributed==1.23.1
Django==2.2
django-contrib-comments==1.9.1
django-shortuuidfield==0.1.3

